+++ .eslintrc.json
@@ -1,3 +1,8 @@
 {
-  "extends": "next/core-web-vitals"
+  "extends": ["next","next/core-web-vitals","prettier"],
+  "rules": {
+    "react/prop-types": "off",
+		"react/react-in-jsx-scope": "off",
+		"@typescript-eslint/no-explicit-any": "off"
+  }
 }
