+++ src/pages/_document.tsx
@@ -1,0 +1,15 @@
+import NextDocument, { Html, Head, Main, NextScript } from "next/document"
+
+export default class Document extends NextDocument {
+  render() {
+    return (
+      <Html lang="jp">
+        <Head></Head>
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    )
+  }
+}
