+++ src/pages/index.tsx
@@ -1,0 +1,18 @@
+import { Box } from "@chakra-ui/react"
+import type { NextPage } from "next"
+import Head from "next/head"
+
+const Home: NextPage = () => {
+	return (
+		<div>
+			<Head>
+				<title>Create Next App</title>
+				<meta name="description" content="Generated by create next app" />
+				<link rel="icon" href="/favicon.ico" />
+			</Head>
+			<Box>Hello,World!</Box>
+		</div>
+	)
+}
+
+export default Home
